Fix the return type of the morton encode64().

+++ PointCloudCompresser/src/MortonCode.cpp
@@ -8,7 +8,7 @@ unsigned int CPC::MortonCode::encode32(const Index & index)
     return libmorton::morton3D_32_encode(index.index.x(), index.index.y(), index.index.z());
 }
 
-unsigned int CPC::MortonCode::encode64(const Index & index)
+unsigned long long CPC::MortonCode::encode64(const Index & index)
 {
     return libmorton::morton3D_64_encode(index.index.x(), index.index.y(), index.index.z());
 }
