+++ PointCloudCompresser/src/MortonCode.h
@@ -8,7 +8,7 @@ namespace CPC
     {
         public:
             static unsigned int encode32(const Index & index);
-            static unsigned int encode64(const Index & index);
+            static unsigned long long encode64(const Index & index);
             static Index decode32(const unsigned int code);
             static Index decode64(const unsigned long long code);    
     };
